### 2016-06-12 #27 @whedon
Submission: sourmash: a library for MinHash sketching of DNA
**Submitting author:** @ctb (<a href="http://orcid.org/0000-0001-6001-2677">C. Titus Brown</a>)
**Repository:** <a href="https://github.com/dib-lab/sourmash/" target ="_blank">https://github.com/dib-lab/sourmash/</a>
**Version:** 1.0

- [x] **Archive:** Does the <a target="_blank" href="https://zenodo.org/badge/latestdoi/22555/dib-lab/sourmash">software archive</a> resolve?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
