### 2016-06-12 5111261
Dockerfile update
