+++ Dockerfile
@@ -1,8 +1,10 @@
+# for binder
 FROM andrewosh/binder-base
 
 USER root
 
 ENV PACKAGES python-dev zlib1g git python-setuptools g++ make ca-certificates
+ENV SOURMASH_VERSION master
 
 ### don't modify things below here for version updates etc.
 
@@ -12,7 +14,9 @@ RUN apt-get update && \
     apt-get install -y --no-install-recommends ${PACKAGES} && \
     apt-get clean
 
-USER main
+RUN cd /home && \
+    git clone https://github.com/dib-lab/sourmash.git -b ${KHMER_VERSION} && \
+    cd sourmash && \
+    python setup.py install
 
-RUN pip install -U setuptools pip
-RUN pip install sourmash
+USER main
