### 2016-06-12 35a25df
more Dockerfile update
