### 2016-06-12 e52056c
even more dockerfile
