### 2016-06-12 87c0b1c
try again to fix Dockerfile
