### 2016-06-15 0cb1872
fix sort bug in 'sourmash search'
