### 2016-06-15 #27 @ctb
Hi @jkahn thx - note that I just updated master with a bug fix (and a test for the bug :) for a bug discovered yesterday.
