### 2016-06-15 9f60602
remove duplicate header stuff in C module
