### 2016-06-16 #27 @jkahn
I am not a copyright lawyer, but I don't think the text given in the release is exactly what [the BSD 3-clause](https://opensource.org/licenses/BSD-3-Clause) would say. For one thing, the Regents of the U of C *and* the MSU are mentioned the `LICENSE` file, but only the RofUofC in the header.  (I understand that @ctb has probably moved from one to the other, looking at @dib-lab .
