### 2016-06-16 dib-lab/sourmash#16 @jkahn
fix license to no longer mention MSU?
see https://github.com/openjournals/joss-reviews/issues/27
