### 2016-06-16 dib-lab/sourmash#18 @jkahn
developer/testing doc needed
@ctb is on it, but I think there needs to be a few pieces of better documentation for how developers can get access to the test environment:
https://github.com/openjournals/joss-reviews/issues/27

For example, #12 and #10 both suggest that there are a few additional steps to take to work as an active developer (rather than a client) of the C bindings.
