### 2016-06-16 dib-lab/sourmash#18 @jkahn
Also, whatever is going on with the `SOURMASH_COVERAGE` environment variable belongs in that doc.
