### 2016-06-16 dib-lab/sourmash#19 @jkahn
test code is in unexpected places
for example:
`sourmash_lib/__init__.py` is an unexpected place to find tests. Do they belong here, or can they be moved to a `tests/test_*.py` constellation?

Similarly, the files in `sourmash_lib/` include test files, which are installed by pip; this is not appropriate for clients who are not developing the `sourmash` library itself.  Probably, all of `sourmash_tst_utils.py` and `sourmash_lib/test_*.py` belong in a `$repo/tests` directory or they need to be renamed.

https://github.com/openjournals/joss-reviews/issues/27
