### 2016-06-16 dib-lab/sourmash#20 @jkahn
paper referenced in .bib needs DOI (here it is)
Please include "doi: http://dx.doi.org/10.1101/029827" into `paper.bib`

https://github.com/openjournals/joss-reviews/issues/27
