### 2016-06-17 dib-lab/sourmash#22 @jkahn
Undocumented utility scripts & directories found at root
... for example, the `urchin`, `utils`, `data`, `refseq` and `sigs` directories.

I could imagine that these directories are any of the following:
- test data to avoid regressions
- useful for particular research questions that `sourmash` is used for
- demo data (for the notebook demos)
or some combination of the above. But as a developer *or* a client of the library, not knowing induces significant cognitive overhead.

I wonder if a separate package of `sourmash_demo` belongs in here somewhere. If it's useful for particular research questions, then maybe there should be addendum or supplementary packages that are bundled separately for PyPi clients or sourmash developers, (depending on who the audience is).

https://github.com/openjournals/joss-reviews/issues/27
