### 2016-06-17 #27 @jkahn
updated the checklist so that there are unchecked boxes corresponding to the open questions.
