### 2016-06-23 d512bd0
fix bug that ignored more than one ksize in sourmash compute
