### 2016-06-23 dib-lab/sourmash#19 @ctb
This is intentional! It can be helpful to have the tests around to debug installations.
