### 2016-06-23 833f036
remove unneeded files
