### 2016-06-23 1a7f0fc
added extras to setup.py
