### 2016-06-27 36c81e9
remove unnecessary files
