### 2016-06-27 355d9b8
cleaned up & removed various data and demo files
