### 2016-06-29 dib-lab/sourmash#16 @ctb
Removed!
