### 2016-06-29 dib-lab/sourmash#22 @ctb
OK - I've moved urchin/ into data/, and removed refseq and sigs. data/ is used in the tests, and utils/ is handy for hacky downloading stuff. I will comment them appropriately.

I'd rather avoid package/repo profileration (check out both dib-lab/ and ctb/ if you want to see why :).
