### 2016-06-29 dib-lab/sourmash#22 @ctb
data/README.rst added in 935d157
