+++ utils/labels.txt
@@ -1,16 +1,0 @@
-SRR2060939_1.fastq.gz mouse/house/11
-SRR2060939_2.fastq.gz mouse/house/12
-SRR2241509_1.fastq.gz mouse/schistosoma/11
-SRR2241509_2.fastq.gz mouse/schistosoma/12
-SRR2255622_1.fastq.gz mouse/schistosoma/21
-SRR453566_1.fastq.gz yeast/x1
-SRR453569_1.fastq.gz yeast/11
-SRR453570_1.fastq.gz yeast/21
-SRR2060939_1.fastq.gz.abundtrim mouse/house/11
-SRR2060939_2.fastq.gz.abundtrim mouse/house/12
-SRR2241509_1.fastq.gz.abundtrim mouse/schistosoma/11
-SRR2241509_2.fastq.gz.abundtrim mouse/schistosoma/12
-SRR2255622_1.fastq.gz.abundtrim mouse/schistosoma/21
-SRR453566_1.fastq.gz.abundtrim yeast/x1
-SRR453569_1.fastq.gz.abundtrim yeast/11
-SRR453570_1.fastq.gz.abundtrim yeast/21
