### 2016-06-30 cf8b698
added README to utils
