### 2016-06-30 dib-lab/sourmash#22 @ctb
utils/README.md added in cf8b6981.
