### 2016-07-10 417572a
fixed "sourmash compute -o" and wrote accompanying test
