+++ doc/developer.rst
@@ -1,0 +1,16 @@
+=====================
+Developer information
+=====================
+
+We use drone.io for continuous integration:
+https://drone.io/github.com/dib-lab/sourmash/
+
+Code coverage calculation is enabled (on Linux only) by running
+'make coverage'.  This recompiles the C++ extension without
+optimization and with coverage configured.  See 'setup.py' for
+more information on this.
+
+Code coverage can be viewed interactively at codecov.io:
+https://codecov.io/gh/dib-lab/sourmash/
+
+
