+++ Makefile
@@ -9,6 +9,12 @@ clean:
 	$(PYTHON) setup.py clean --all
 	cd doc && make clean
 
+install:
+	$(PYTHON) setup.py install
+
+dist: FORCE
+	$(PYTHON) setup.py sdist
+
 test: all
 	$(PYTHON) -m pytest
 
@@ -19,3 +25,5 @@ coverage: all
 	$(PYTHON) setup.py clean --all
 	SOURMASH_COVERAGE=1 $(PYTHON) setup.py build_ext -i
 	$(PYTHON) -m pytest --cov=.
+
+FORCE:
