+++ MANIFEST.in
@@ -1,1 +1,7 @@
-recursive-include third-party *.cc *.h
+include LICENSE
+recursive-include sourmash_lib *
+recursive-include third-party *.cc *.hh
+recursive-include lib *.hh *.cc [Mm]akefile* *.rst README*
+recursive-include README.rst
+global-exclude *.orig
+global-exclude *.pyc
