### 2016-07-10 0c4e91c
include numpy etc in test extras
