+++ doc/release.rst
@@ -17,7 +17,7 @@ Testing a release
 #. Set your new version number and release candidate::
 
         new_version=0.9.3
-        rc=rc3
+        rc=rc5
 
    and then tag the release candidate with the new version number prefixed by
    the letter 'v'::
