+++ MANIFEST.in
@@ -1,4 +1,5 @@
-include LICENSE Makefile Dockerfile LICENSE Makefile README.md
+include LICENSE Makefile Dockerfile LICENSE Makefile README.md requirements.txt
+include index.ipynb
 recursive-include sourmash_lib *
 recursive-include third-party *.cc *.h
 recursive-include README.rst
