### 2016-07-11 d4fb9d5
add dev docs; update release insts
