+++ doc/developer.rst
@@ -2,6 +2,30 @@
 Developer information
 =====================
 
+Development environment
+-----------------------
+
+You can get the latest development master branch with::
+
+    git clone https://github.com/dib-lab/sourmash.git
+
+To install all of the necessary dependencies, do::
+
+    pip install -r requirements.txt
+
+Briefly, we use ``py.test`` for testing, and ``coverage`` for code
+coverage analysis.
+
+We suggest working on sourmash in a virtualenv; e.g. from within the
+sourmash clone directory, you can do::
+
+    python -m virtualenv dev
+    . dev/bin/activate
+    make clean all test
+
+Automated tests and code coverage calculation
+---------------------------------------------
+
 We use drone.io for continuous integration:
 https://drone.io/github.com/dib-lab/sourmash/
 
