### 2016-07-11 dib-lab/sourmash#18 @ctb
Mostly done, at https://github.com/dib-lab/sourmash/blob/master/doc/developer.rst
