### 2016-07-12 0a946f7
add binder comment to release doc
