### 2016-07-12 0143d52
update test discovery
