+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 addopts = --doctest-glob='doc/*.rst'
 python_files = sourmash_lib/*.py
-norecursedirs = utils
+norecursedirs = utils build buildenv
