+++ paper.bib
@@ -5,4 +5,6 @@
   pages={029827},
   year={2015},
   publisher={Cold Spring Harbor Labs Journals}
+  doi={10.1101/029827}
+  url={http://dx.doi.org/10.1101/029827}
 }
