### 2016-07-12 dib-lab/sourmash#20 @ctb
Added in 944a6d6
