### 2016-07-12 52cc1bf
add SOURMASH_COVERAGE info
