+++ doc/developer.rst
@@ -32,7 +32,9 @@ https://drone.io/github.com/dib-lab/sourmash/
 Code coverage calculation is enabled (on Linux only) by running
 'make coverage'.  This recompiles the C++ extension without
 optimization and with coverage configured.  See 'setup.py' for
-more information on this.
+more information on this; the environment variable
+``SOURMASH_COVERAGE`` controls whether the C++ extension is
+compiled with code coverage analysis enabled.
 
 Code coverage can be viewed interactively at codecov.io:
 https://codecov.io/gh/dib-lab/sourmash/
