### 2016-07-12 dib-lab/sourmash#18 @ctb
Finished, in 52cc1bf
