### 2016-07-12 4c689ee
created entrypoint script
