+++ sourmash_lib/__main__.py
@@ -1,0 +1,341 @@
+from __future__ import print_function
+import sys
+import os, os.path
+import argparse
+import csv
+
+import screed
+import sourmash_lib
+from sourmash_lib import signature as sig
+from sourmash_lib import fig as sourmash_fig
+
+DEFAULT_K = 31
+DEFAULT_N = 500
+
+
+class SourmashCommands(object):
+
+    def __init__(self):
+        parser = argparse.ArgumentParser(description='work with RNAseq signatures',
+                                         usage='''sourmash <command> [<args>]
+
+Commands can be:
+
+   compute <filenames>         Compute signatures for sequences in these files.
+   compare <filenames.sig>     Compute distance matrix for given signatures.
+   search <query> <against>    Search for matching signatures.
+   plot <matrix>               Plot a distance matrix made by 'compare'.
+
+   import_csv                  Import signatures from a CSV file.
+.
+''')
+        parser.add_argument('command')
+        args = parser.parse_args(sys.argv[1:2])
+        if not hasattr(self, args.command):
+            print('Unrecognized command')
+            parser.print_help()
+            sys.exit(1)
+
+        cmd = getattr(self, args.command)
+        print('# running sourmash subcommand: %s' % args.command,
+              file=sys.stderr)
+        cmd(sys.argv[2:])
+
+    def search(self, args):
+        "Search a query sig against one or more signatures; report top match."
+        parser = argparse.ArgumentParser()
+        parser.add_argument('query')
+        parser.add_argument('against', nargs='+')
+        parser.add_argument('--threshold', default=0.08, type=float)
+        parser.add_argument('-k', '--ksize', default=DEFAULT_K, type=int)
+        parser.add_argument('-f', '--force', action='store_true')
+        args = parser.parse_args(args)
+
+        # get the query signature
+        sl = sig.load_signatures(
+            open(args.query, 'r'), select_ksize=args.ksize)
+        if len(sl) != 1:
+            raise Exception("%d query signatures; need exactly one" % len(sl))
+        query = sl[0]
+
+        # get the signatures to query
+        print('loading db of signatures from %d files' % len(args.against),
+              file=sys.stderr)
+        against = []
+        for filename in args.against:
+            if filename == args.query and not args.force:
+                print('excluding query from database (file %s)' % filename,
+                      file=sys.stderr)
+                continue
+
+            sl = sig.load_signatures(
+                open(filename, 'r'), select_ksize=args.ksize)
+            for x in sl:
+                against.append((x, filename))
+
+        # compute query x db
+        distances = []
+        for (x, filename) in against:
+            distance = query.similarity(x)
+            if distance >= args.threshold:
+                distances.append((distance, x, filename))
+
+        # any matches? sort, show.
+        if distances:
+            distances.sort(reverse=True, key = lambda x: x[0])
+            print('%d matches:' % len(distances))
+            for distance, match, filename in distances[:3]:
+                print('\t', match.name(), '\t', "%.3f" % distance,
+                      '\t', filename)
+        else:
+            print('** no matches in %d signatures' % len(against),
+                  file=sys.stderr)
+
+    def compute(self, args):
+        "Compute the signature for one or more files."
+        parser = argparse.ArgumentParser()
+        parser.add_argument('filenames', nargs='+')
+        parser.add_argument('--protein', action='store_true')
+        parser.add_argument('--input-is-protein', action='store_true')
+        parser.add_argument('-k', '--ksizes',
+                            default=str(DEFAULT_K),
+                            help='comma-separated list of k-mer sizes')
[…]
