### 2016-07-12 bc77667
remove 'sourmash' script from dist
