+++ setup.py
@@ -66,7 +66,6 @@ SETUP_METADATA = \
                               language="c++",
                               extra_compile_args=EXTRA_COMPILE_ARGS,
                               extra_link_args=EXTRA_LINK_ARGS)],
-    "scripts": ["sourmash"],
     "install_requires": ["screed>=0.9", "PyYAML>=3.11"],
     "extras_require": {
         'test' : ['pytest', 'pytest-cov', 'numpy', 'matplotlib', 'scipy'],
