### 2016-07-12 2ed0122
fixed setup.py
