### 2016-07-12 408dc6a
explicitly set success return code
