### 2016-07-12 #27 @ctb
Hi @jkahn Thanks for all your comments - many excellent improvements resulted. I think this is ready for review once again!
