### 2016-07-12 #27 @ctb
Note, the latest version is 0.9.4, https://github.com/dib-lab/sourmash/releases/tag/v0.9.4, and zenodo etc should be updated soon
