### 2016-07-18 dib-lab/sourmash#28 @jkahn
Makefile install needs to specify test extras
The following pattern should work, but draws errors:

```
git checkout v0.9.4
mkvirtualenv --python=/usr/bin/python2.7 sourmash-py27
make clean install test
```
But if the `install` target was using `pip install .[test]` this *would* work. I've not quite got the PR formatted, but I'll just file the ticket here, as part of my second-pass review for [JOSS](https://github.com/openjournals/joss-reviews/issues/27).
