### 2016-07-18 dib-lab/sourmash#30 @jkahn
Developer Makefile incomplete WRT sphinx, pip install
I did:
```
mkvirtualenv sourmash-py27
pip install -r requirements.txt # implied but not specified in docs
make clean all test
```
this drew a failure because `sphinx` wasn't installed.

Starting fresh:
```
pip install -r requirements.txt sphinx
make clean all test
```
succeeds (if you don't have a `.tox` directory lying around messing things up).

I might add that `all` is a funny name for a build target that builds the C++ library.

Minor concerns but still needs to be addressed for [ease of developer entree](https://github.com/openjournals/joss-reviews/issues/27).
