### 2016-07-18 #27 @jkahn
Pending these two concerns (both resolved by a little bit more attention to onboarding new developers), I'm satisfied that my concerns have been addressed and recommend the package re-mint and be accepted.

I want to address my one continuing gripe: I don't love that [the test code is mixed in (closed, wontfix)](https://github.com/dib-lab/sourmash/issues/19) among the shipped code (in a public-facing library, this could open an entirely new attack surface) but I don't think it's a release-blocking offense, especially in a research package.
