### 2016-07-22 #27 @arfon
> Pending these two concerns (both resolved by a little bit more attention to onboarding new developers), I'm satisfied that my concerns have been addressed and recommend the package re-mint and be accepted.

👍 thanks @jkahn. @ctb - please let me know when you've made these changes.
