### 2016-09-05 baac87b
add rkmh to docs; fix link in md format
