### 2016-09-11 5aaec5c
remove 'test.sh' from tox.ini
