+++ tox.ini
@@ -8,5 +8,4 @@ deps=
     screed
 	pytest
 commands=
-    ./test.sh
 	make test
