### 2016-09-11 64597b5
update make test to install test extras
