### 2016-09-11 dib-lab/sourmash#28 @ctb
`make test` now does `pip install '.[test]'`, as of 64597b.

Note that having the install target install the test extras would defeat the purpose of having them be extras - might as well just have them be default :). Admittedly, pretty much anyone using the Makefile is probably going to be a developer, so it probably doesn't matter.
