### 2016-09-11 3f0e49a
remove 'make all' from dev instructions
