### 2016-09-11 dib-lab/sourmash#30 @ctb
The commands:

```
mkvirtualenv tst
make clean test
```

work fine as of #28 fix. In 3f0e49a, I've updated the developer documentation to remove the 'all' target since it's unnecessary with 'test'.

I've always seen & used `all` as the name of the default target in a Makefile. I'm not sure why it's odd in this contest?
