### 2016-09-11 dib-lab/sourmash#30 @ctb
(thanks for the bug report!)
