### 2016-09-12 83a1273
fix a few minor issues in copy/paste commands
