### 2016-09-13 #27 @ctb
OK, I've made the changes and @luizirber has checked them twice :).

Am I correct in thinking that I should proceed by cutting a new release (v1.0) and then putting the zenodo link in here? That's it, yes? (Reading [author guidelines](http://joss.theoj.org/about#author_guidelines))
