### 2016-09-13 #27 @arfon
> Am I correct in thinking that I should proceed by cutting a new release (v1.0) and then putting the zenodo link in here? That's it, yes? (Reading author guidelines)

👍 yes please @ctb
