### 2016-09-13 a874442
update for v1.0
