### 2016-09-14 #27 @ctb
[sourmash 1.0](https://github.com/dib-lab/sourmash/releases/tag/v1.0) now up! [![DOI](https://zenodo.org/badge/55858065.svg)](https://zenodo.org/badge/latestdoi/55858065)
