### 2016-09-14 #27 @arfon
Many thanks again for the review @jkahn!

@ctb your paper is now accepted in JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00027 🎉 🚀 💥
